Change the default value of `states_required` to false
The default value of states_required column of :spree_countries table is
true which makes the configuration option
Spree::Config.address_requires_state essentially a no-op. Therefore,
setting the aforementioned configuration option to false is not enough
to remove the state requirement of a country.

This commit adds a migration for the :spree_countries table with the
name ChangeStatesRequiredForCountries which changes the default value of
states_required to false.

As the states_required value is now false, this value has to be set
explicitly to true for federated states. Using the list of federated
states from http://en.wikipedia.org/wiki/Federated_state set the
states_required value of these states explicitly to true.

Fixes: #3709

Signed-off-by: Ali Polatel <ali.polatel@ozguryazilim.com.tr>